requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
icalendar
pytz
python-dotenv
urllib3<2.0.0